requirements.txt
absl-py==0.13.0
numpy==1.19.5
tensorflow>=2.1.0
